Add variables for managing under/overcloud update/upgrade
This change introduces variables to enable upgrade or update of
the underloud/overcloud respectively. By default these are set
to false and the scripts for running the upgrade/update commands
only get created but not run.
It also adds a main.yml file which is required by Infrared to run
this role as a plugin which maps the options defined in plugin.spec
to the role variables.

Change-Id: I9ed0f9aeb8bc37be6443e684a3d8009eeb993ab6