Use parameter to control the docker registry env file creation
By setting force: no the created script files are rendered only
once. The role should allow multiple runs and update the scripts
according to the latest options passed by the user. This change
controls the scripts creation by vars in order to be able to
create or not the script files.

Change-Id: I1594ddfda11ac638ba65625964c0a492ad9e6e18