Add condition to create container images download script.
It is also interesting for the role to support
the option of passing by an already existing
script that handles the container images downloading.

By adding this conditional we avoid rewritting the
content of the passed script when setting the var
to false.

Change-Id: I8b82071dfa7d440c7c840ffb9bcdf006843b6d97
(cherry picked from commit 6377a43a1721404ba32cf558e3493d3018c97ab3)