More robust check for deployment_with_ha
The current check of "-pacemaker" being present in deployment command
is not sufficient for OOOQ based environments (and CI). Checking the
EnabledServices stack output should be somewhat more robust. We only
want deployment_with_ha if the deployment was with
pacemaker (pacemaker_present test) and we can't assume pacemaker is
the default (is_upstream test).

Co-Authored-By: Jose Luis Franco <jfrancoa@redhat.com>
Change-Id: I8b0ff3797d8f6d697086701c80e1ae159cff2bf4