Less opinionated detection of service env files
This should support OOOQ while hopefully not damaging Infrared
integration. We need to specifically look for the env files that match
the patterns we want, and we also need to support multiple env files
on a single line, because that's how OOOQ passes them in.

This solution should work both for the transformed and untransformed
deployment commands.

Co-Authored-By: Jose Luis Franco <jfrancoa@redhat.com>
Change-Id: I4164ef552c333d6b9672c1a3d55c15a1b76e7bb7