Run compute upgrade only in compute hosts.
The basic deployment used in TripleO CI
with only a controller is displaying
the controller in the openstack hypervisor
list command and tries running
upgrade-non-controller.sh on it, which fails.

With this change we plan to filter out
only those hosts with the compute role, leaving
aside those hosts in which nova-scheduler and
nova-compute are running at the same time.

Closes-Bug: #1747432
Change-Id: If7bfd63d62509387be4b843d470816365f68cfaf