Add log artifacts for upgrade steps
Context: https://trello.com/c/vJwmd6SD/771-indexing-upgrade-logs-in-logstash
         https://trello.com/c/TU3JEtmg/560-ensure-upgrade-logging

- Per feedback have added timestamper_cmd so these
  logs can be indexed by logstash.

- for clarity have added a distinct variable for update template

- for clarity have added ffu_ prefix to FFU variant of UC upgrade template

- Per feedback removed the logging from bash jinja template
  as it's inconsistent with logging norms in TQ/TQE

I suggest that we also rename undercloud_upgrade.sh.j2 to reflect
that it's used in both upgrade and update workflows, or we clearly
disambiguate and fully denormalize by duplicating the template.  In order
to limit scope/risk IMHO this should be done in a discreet commit.

Change-Id: I184c5ebd28ae11a93a854df5dc2cafd02a5e1b6c