Replace scenario file in oooq ffu upgrade.
The scenario file during upgrade must match
upgrade_composable_scenario.

In oooq, composable_scenario variable is used for deployment and
upgrade_composable_scenario is used for upgrade.  So we need a way to
adjust the script to the upgrade_composable_scenario file.

For upgrade it's done there[1] for upgrade.  I think that for the
proposed way for FFU is more generic.

[1] https://github.com/openstack/tripleo-upgrade/blob/master/tasks/upgrade/create-upgrade-scripts.yaml#L64..L70

Change-Id: I5bfe7e56bc5c8d3d3f9d0a7f8112372b88fc725c