New major upgrade workflow implementation.
This patch includes the new upgrades workflow to
upgrade to Queens into the tripleo-upgrade role.
A overcloud_upgrade_prepare.sh script will be run
to get the HEAT outputs into Ansible playbooks which
will be run later via as many scripts as roles in
the environment (Controller, Compute, etc...).
In each of these scripts the command to run the
downloaded playbooks is executed.

All containers preparation and downloading tasks
have been moved out of the upgrade playbook, so
it becomes a pre-requisite to have uploaded the
right containers before upgrading using this role.

All this new workflow is based on the steps defined
in [0].

[0] https://review.openstack.org/#/c/535859/

Change-Id: Iae680d97eeef52b9145c290b0b988cb267af3b3d
(cherry picked from commit 799eea86d53577be885791cfb3f3e191420d7b65)