Include connectivity check prepare scripts during FFU
This change includes the connectivity check prepare scripts step
during FFU and it moves the tasks covering it to the
common dir and assigns specific tags for each FFU stage. In
addition it saves the floating IP address to a file so it can be
used in stages where Neutron API is not available.

Change-Id: I9d61f6eaa105e9eb2854eb749f570c395de155fe
(cherry picked from commit 55ff3f72cdf8d9acc15c22a6ee72a215fc324fcc)