Add FFU converge step
https://review.openstack.org/#/c/559061/ introduced a ffwd-upgrade-
converge.yaml environment file to be used at the end of the ffu
process, before running the ceph upgrade. This change adjusts the
workflow to include the converge stage.

In addition it adjusts the scenario environment file in the overcloud
deploy script so any scripts built of it make use of the new one.

Change-Id: I1b2c9e0d804b511c6c6e2cce34ed83fca8065bae