Clean up ansible variables.
The amount of variables defined inside the
defaults/main.yml was piling up. It makes
sense to provide flexibility via variables,
but some of the ones defined in the role
didn't change and won't change during its
use.

This patch remove most of the '_template'
variables and add the template name in the
src section of the task. Remove also workload variables.

Change-Id: I443865bc95590a5cffcc28d98913110238e863b6