Allow running update/upgrade converge with --update-plan-only
We discovered we need to put back stack updates into the converge
operations. This increases the run time of the upgrade procedure
considerably, and e.g. our voting CI job for updates might go over the
time limit. We'll need to start passing --update-plan-only to the
converge operation to keep CI on previous run times.

Depends-On: Idc5dbebc116aba2420d275aab3981186511f298e
Change-Id: I9039db99f43e8ad091ea6fef7e3a149fcddf1334
Related-Bug: #1768586