Separate scripts creation tasks for under/overcloud.
The upgrade/update tasks in tripleo-upgrade started
with a create-scripts book, which generated most of
the scripts from the templates. However, this book
mixed the creation of scripts needed for the overcloud
as well as for the undercloud. This becomes an issue
when we want to perform only an undercloud upgrade, as
we are running tasks intended to the overcloud and it
might not even exist.
This patch divides the script creation tasks used for
the undercloud from the ones used for the overcloud.

Change-Id: I35dfbe50d12b65a8d14f62410adea551e7584912
Closes-Bug: #1769107
(cherry picked from commit 0fd5d40994cb30117444b22a9e862722c4ebc52c)