Separate scripts creation tasks for under/overcloud.
The upgrade/update tasks in tripleo-upgrade started
with a create-scripts book, which generated most of
the scripts from the templates. However, this book
mixed the creation of scripts needed for the overcloud
as well as for the undercloud. This becomes an issue
when we want to perform only an undercloud upgrade, as
we are running tasks intended to the overcloud and it
might not even exist.
This patch divides the script creation tasks used for
the undercloud from the ones used for the overcloud.

Change-Id: I7e6c288dd0f974901fb85acb1ac806bcd4d9ee4d
Closes-Bug: #1769103
(cherry picked from commit abc439355ee433db4a1e21c8da31d8f48b7b03e5)