Add splitstack detection to auxilary-facts
The upgrade prepare/converge templates are used for ffu as well. This
change runs the steps for detecting if a splitstack env is in use
in the auxiliary-facts so the splitstack_deployment var is made
available for ffu as well.

Change-Id: I49a34e323cf9479ef387e20b7f831d574a724cd8
(cherry picked from commit 91f7034f66db7596079b80c0e8d8bca35695710e)