Change include for include_tasks when including loop controled scripts.
We obtain a 'controllers' is undefined ansible error when
the tripleo-upgrade is run without any role related tag
(e.g. overcloud_upgrade, overcloud_update, etc...). The
issue seems to appear at the time of including the
controller_post_script.yml tasks, which content is defined
depending on the result of a shell command. As the command
is not executed during the role importing, the include fails.

The new include_tasks keyword seems to handle this issue
without displaying any error. Also the include keyword was
deprecated in Ansible 2.4, so it's time to start moving into
include_tasks/import_tasks.

Change-Id: I9d10e1d9a95e11fd59bf593bb052abf9eb17cf78
Closes-Bug: #1783303