Adjust roles_data also for updates.
During the update prepare process, if
some of the services defined in the roles
data where deprecated in the version to
update to, an error will be retrieved during
container image downloading as the service
is not present anymore. This patch enables
the roles_data adjusting, which was run in
upgrades, also in updates.

Change-Id: I1e1fc6471525855fb09280131e7847f21b1a74d8
Related-Bug: #1784624
(cherry picked from commit ba9879a93209ba25fdb9a5dd1d00aa3a5f33a9bf)