Don't invoke multiple migration requests for single vm.
This is Pike only fix. If there are multiple migration requests
for the same instance, the placement inventory might
not be cleaned up in time, leading to further migration requests
to this compute to fail:
  <snip>
  Allocation for VCPU on resource provider
  b6d1973f-bc39-4ea0-8d28-15f988f762e1 violates min_unit, max_unit,
    or step_size. Requested: 5, min_unit: 1, max_unit: 4, step_size: 1
  Placement API returning an error response:
   Unable to allocate inventory:
    Unable to create allocation for 'VCPU' on resource provider
     The requested amount would violate inventory constraints

Change migration timeout to 180 seconds.
Both nova-compute and neutron-openvswitch-agent have to be up to
have live-migration working. Added check for neutron's part.

Closes-Bug: 1785568
Change-Id: If46c43aee9a409b59f2dcb139d739106bc9a562f