Enhance undercloud reboot procedure for baremetal
Rebooting bm undercloud takes longer than vm so procedure has
to be adjusted.

Current mechanism of asserting if system is up is not reliable.
This patch changes it to call wait_for module in do..untill loop:
 try to connect to server to get 'OpenSSH' string with a 30 seconds
timeout, retrying this for 21 time. This results in
 30x21 = 630 sec for whole operation. This might look too high
but better to wait 10 minutes rather than fail job due to slow hw.

Change-Id: Ic463acc7bed84d23033d2b8c00336f1024fe0e3b