Use Jinja2 macro to remove redundancy in workarounds.sh.j2.
Currenty, a lot of code was being repeted in the template which
would force to modify it in multiple places if a change was being
done in one of the workaround blocks. By using the Jinja2 macro,
which behaves as a function, we can reuse the code defining it in
one single place and calling to it multiple times.

Change-Id: I63db9e8b30c9f9e99c501300982b513b627bed07