Add timeout to openstack command
There are cases when udercloud is not ready. This causes 'openstack' cli to
hang which is not good in case of ansible check. This patch adds timeout to
kill the process, then will cause ansible to retry the command again instead of
hang.

Change-Id: Ifaeff772a0c95ddb4d39c0a7b6fdd56a6abb023b
(cherry picked from commit 7e0412eb48e59036872d24544b5a61398eb740ce)