Do not adjust roles data for minor updates.
During a queens minor update, the roles_data already contains the
ContainerImagePrepare service, so we end up appending it twice
into the Controller's role. This caused some errors during minor update.
The custom roles adjustment is not needed anymore, as there were
no services chaged during the rocky release, so we are removing it.

Change-Id: I9afbb421b8dd877c73b6fc29a02f2486e7d5cfa9