Option to specify python version and use package instead of yum.
In rhel8 the python-tripleoclient version changes into python3-
tripleoclient, so to avoid issues with the packages version a
py_ver variable has been added to allow us make use of one
version or the other. Also, do not depend on the yum command, as
in rhel8 there is no importable yum module. We'll use the package
task instead.

Change-Id: I5edf7b63acaa4df15de5b3bacbc7b9cae1e99b5b