Add new oc_hosts_roles variable.
With the new upgrades workflow we need to perform a more
granular upgrade node by node, instead of by role as we
were doing before. For that reason, we're adding a new
variable which contains all the hosts an ansible group
consists of. So we can iterate over them at the time of
upgrading.

Change-Id: I81f4add06218ad201d4772821723154c33d4123e