Upgrade undercloud.conf with the right parameters
This patch adds the new parameters required for the OSP15 major
upgrades workflow.  This patch also adds the undercloud_hiera file
if the hieradata_override is not defined in undercloud.conf, if it
is then it will append the needed hiera configuration into the existing
file.
If the docker_insecure_registries option exists in the file, it's
value is merged with new namespace values and set as
container_insecure_registries.  The docker_insecure_registries is
removed after the value is used.

Change-Id: Ib83e508950df729901fffa072402b581be95c746
(cherry picked from commit 3a97b561fa3becea25d0c2385850c1139358228d)