Add support for network-data-file.
TripleO offers a default network topology when deploying
with network isolation enabled, and this is reflected in
the network_data.yaml file in tripleo-heat-templates.
If this is the case it has to be handled during update/upgrade.

Change-Id: If50109753d4845357c869986c256186e8e22006d
(cherry picked from commit 663d380f6fe0e21d58de3aa55cce3cdf9c849ef1)