Move undercloud upgrade pre-requisites into tasks.
Up to now we were keeping the packages updates needed before running the
undercloud upgrade inside the template templates/undercloud_upgrade.sh.j2.
However, as we need to include new pre-requisites, plus the fact that for
the downstream yum is not present (it's dnf instead) and upstream we still
have yum it's better to ensure these pre-requisites leveraging the benefits
the Ansible tasks bring.

The new set of tasks that will take care of updating those packages and
ensuring the needed services are enabled will be in
common/undercloud_prerequisites.yaml as we'll need them for both the
undercloud upgrade and update.

Also, this patch removes the py_ver variable and relies on use_oooq variable
to differentiate between the python version we need for the packages.

Change-Id: I40b17910ecd963d50a2074081d6dea285d8c9dcc
(cherry picked from commit 999e1803a70383796e6b8465c9405d0b6506d716)