Update tripleo-validations before upgrading the undercloud.
From rocky to stein, the tripleo-validations project has experienced
a big refactoring having some directories and files moved. When
upgrading the undercloud, the openstack-tripleo-validations is not
included as a dependency in the tripleo-heat-templates, tripleo-common
or tripleoclient spec file and as a consequence, the package won't get
updated.

The result is that the undercloud upgrade operation fails because it
can't locate the tripleo-validations files in the right location:
No such playbook: /usr/share/openstack-tripleo-validations/playbooks/
undercloud-disk-space-pre-upgrade.yaml

This patch ensures that openstack-tripleo-validations package is also
upgraded before upgrading the undercloud.

Change-Id: Ibfaa563a5882af0bff33b0e0cba92cb1b6a282fb
Closes-Bug: #1831041