Remove check for openvswitch package.
The package task does ensure that a package is installed,
when the intention of this task was to simply check if
the openvswitch package was installed. However, if it's
installed, Ansible will try to install the latest available
version and that update needs to be handled in the Undercloud
upgrade itself, not in the previous steps. For that reason,
this patch removes the task and relies on the following one
which checks that the OpenVSwitch service is running to fail
if the package wasn't installed.

Change-Id: I542b82b0bec05faf320861027f2f51572fe931c2
(cherry picked from commit b5f00e5c3e663b0ebcfe3277e38a71a3bf91e7d9)