Pass neutron-ovs.yaml environment in upgrade prepare.
If the environmet is deployed in Rocky with OVS then we need to pass
the neutron-ovs.yml environment file for the upgrade prepare and converge
steps to work. Otherwise, as OVN is the Neutron engine by default in
Stein it will try to deploy with OVN and the upgrade will fail.

Also, the containers-prepare-parameter.yaml is shared between the UC
and the OC upgrade. For the case of the UC, there is no possibility to
deploy with ovn, so we force the neutron_driver option (defaulted to ovn
when creating the default containers-prepare-parameter.yaml file) to
openvswitch. However, for the OC, we need to identify if it was deployed
with the OVN service enabled and if so, set the netron_driver to ovn instead.

In summary, we first upgrade from Rocky with OVS to Stein with OVS and
if required, a post-upgrade step to upgrade from OVS to OVN would be
performed. In the case the Rocky deployment was with OVN enabled we
perform a Rocky with OVN to Stein with OVN upgrade.

Change-Id: I21bbf9cda977a1c269e2c1caad4bf88d3f6173b0
(cherry picked from commit bb44ee55a3057540ffd05a4b55f8ccff9294f3e0)