Enable system upgrade and upgrade run per host.
Allow the execution of the overcloud upgrade steps by hosts
withing the Overcloud roles. After this patch is merged, the hosts
contained in the ovecloud roles (Controllers, Computes, Storage, etc...)
will be taken serially to apply the upgrade host by host.

For each of the hosts within these roles we'll execute the following
steps:

1- Pre overcloud system upgrade workarounds (Workaround executed in the
   host via Ansible with --limit)
2a- Overcloud system upgrade prepare + overcloud system upgrade run (if
   system_upgrade_compact == false)
2b- Overcloud system upgrade (if system_upgrade_compact == true)[Default]
3- Overcloud system upgrade transfer (if role == Controller and
   host == first)
4- Post overcloud system upgrade workaronuds (Workaround executed in the
   host via Ansible with --limit)
5- Pre overcloud upgrade run workarounds (Workaround executed in the
   host via Ansible with --limit)
6- Overcloud upgrade run
7- Post overcloud upgrade run workarounds (Workaround executed in the
   host via Ansible with --limit)

These sets of steps will be executed for all the hosts belonging to the
Overcloud Ansible group.

Change-Id: Ia13f479d58979561fca1c274ac7dde5dc8c2343a