Add else statement when running leapp without SKIP_DEVEL.
If leapp_unsubscribed is set to false, the condition won't be
met and the environment variable won't be included. However,
Ansible complains that there is no else statement and the task
fails: the inline if-expression on line 2 evaluated to false and n
o else section was defined. Adding the else statemnt to include
an empty string to solve the issue.

Change-Id: I8a52da5ca81e61a9dd5f39997535965a75a5abba
(cherry picked from commit 4e305799600e104092087f566a15907a348b58e5)