Ensure pacemaker bootstrap controller is upgraded first.
The variable oc_roles_hosts contains all the hosts per role
(Controller, Compute, etc...) retrieved from the Ansble inventory.
However, when building the variable it was not taken into
consideration that the keys() function won't have to return
the keys in an ordered way. This made us realize that if we
want the upgrade to work, we need to ensure that the very first
controller which we upgrade matches with the pacemaker bootstrap
node (value stored in the pacemaker_short_bootstrap_node_name
hiera data variable).

This patch will retrieve the value of that host and make sure
that it's being stored in the first place for the hosts which
belong to the Controller role.

Change-Id: Ica801b27d09c56fbc1ce7a970c004cd7a5c7b7bf