Fix image validation when AP services get several images.
When AP (Actif/Passif) service get several images back from the host
the script is broken because it wrongly tests any left over image.

We fix that by sorting them by tag (the higher being at bottom of the
list) and taking only the last one.

We ensure as well that re-running the script doesn't unnecessarily
increase the number of hosts we test for AP services.

Change-Id: I24e05a15975fb5a696d195d87257bc7b7e1c561b
Related-Bug: #1843023