Handle upgrade_prepare_extra_params as comma-separated string.
Infrared handles the extra-vars parameters passed as a list as strings.
Currently, tripleo-upgrade differentiates if the parameter is a string
or a list, but it does not cover the case when the parameter was a list
casted into a string. The solution implemented is to accept a comma
separated string like -e upgrade_prepare_extra_params="/home/stack/
overcloud-params.yaml,/home/stack/overcloud-repos.yaml" which then
will be passed into the overcloud upgrade prepare.

Change-Id: Iec0cd44ee79708d48f9c40b58e0a9cbed76d3e9c