Do not validate neutron related side car container images.
Neutron spawns side car container when it needs them. In stein that
would be neutron-haproxy-ovnmeta which has an haproxy that relays
metadata queries.

When we run an update those side car container are not updated to not
disrupt services, so we add them to the list of container to not
check.

Change-Id: Ic1f46e1978bbb34f80aed8a761c38acedd040296