Update cirros image URL to use HTTPS.
The launch_workload test downloads a cirros image to
deploy a small set of instances in the overcloud.
However, the URL being used up to now is not available
any more as it's required to use secure HTTP. This
patch needs to be cherry-picked into all the releases
or it won't be possible to launch the check.

Change-Id: I676225dcfd18ac8423fb802d93bddffdeb29ef5f