Do not remove rhos-release if running leapp unsubscribed.
The option leapp_unsubscribed allows us to run the Leapp
upgrade without need of a subscription (relies on external
workarounds and hacks). As it uses rhos-release underneath
to provide with the RHEL8 repos, we can't get rid of rhos-
release as we were doing when the node was subscribed.

Change-Id: I08abf241ee742622593e5c841db127fcf78c3d13
(cherry picked from commit b66686c706f1beb93d07e7ae929dc844fa2948af)