Fix python command and fix python3 compatiblity.
The python binary is not available in rhel8, add a variable with the
binary. In stein on, we deal python3 by default and print isn't
supported anymore now print is a function. This patch defines python_bin
based on OS version (python in RHEL7 or python3 in RHEL8). Output is
assigned as ansible registered variable to save it to file later on.
Additionally this patch moves all pipes to jq to minimize number of forks

Co-Authored-By: Sergii Golovatiuk <sgolovat@redhat.com>
Closes-Bug: #1856865

Change-Id: I8b2cacd4271a59dfda948462146c0866b8b7725f