Run upgrade for previous hosts only in Controller role.
The upgrade from Rocky to Stein is quite special as it
needs to handle the OS upgrade. In the case of the Controllers
we create progressively a new cluster with the upgraded nodes.
Therefore we need to pass the already upgraded nodes everytime
we upgrade one (pass controller-0 and controller-1 when running
the upgrade of controller-1). However, this is not needed for
any other role and the current logic was implementing it for
all. The result was that we were running the upgrade of the
first compute twice (first compute-0 upgrade and second time when
running upgrade for compute-0 and compute-1).

This patch limits the HA dance only to the Controller role.

Change-Id: I864f3f2ca90416f567f94bff34a34bdbcea9237e
(cherry picked from commit 9ad387cb783409c7ccc0e08e25b005d43081d6ab)