Change roles_data/update_serial without using shell
The currently mechanism uses shell/command, cat and jq in
4 tasks.

We can do the required transformation without using shell,
command, cat or jq by using the slurp ansible module and
some jinja. We can also reduce the implementation to 3
tasks.

This patch implements that in order for a more optimal
execution which is also easier to read and maintain.

Related-Bug: #1856865
Change-Id: Iecf2b0211a69e99efe0f402efb00c7080b784104