Force OS_COMPUTE_API_VERSION to a lower one on workload test.
After the upgrade prepare and right before the upgrade run
command a workload test is executed, which creates an instance
and performs a pingtest during the whole upgrade procedure.
The generated rc file for the OC has the OS_COMPUTE_API_VERSION
pointing to 2.latest which provokes a conflict in this situation
as we have a OSP16 Undercloud and a OSP15 overcloud, failing with
the error: Version 2.79 is not supported by the API. Minimum is
Minimum is 2.1 and maximum is 2.72.

This patch forces the maximum version on the script launch to be
2.72 at max.

Change-Id: I8c949de2083fa6b251ad7c8c6a2e57d6b9c3465e