Allow to switch from parallel to serial update.
We can encounter corner case pacemaker issues with parallel role
update. While we solve them, we need a way to disable parallel role
update.

Using a idiom mentionned in the ansible documentation[1] we start role
update by batch. When batch is 1, this is serial update, one role
after another.

This is the default.

[1] https://docs.ansible.com/ansible/latest/user_guide/playbooks_async.html

Change-Id: I03378557653d07113fa70782e5d22bf5e3e969b8