Switch from using cat to slurp
Slurp is binary safe, ensuring that all contents of the file
are never interpreted or changed during ingestion or transit.

As a preventative measure we should always use it rather than
assume that using cat will result in safe content transit.

Note that tasks/common/adjust-roles-data.yaml is left as-is
due to the change to use slurp already being included in
Iecf2b0211a69e99efe0f402efb00c7080b784104

Change-Id: Ib8309dca36ac804ec17761f6301b1b220411afff