Ensure oc_roles and oc_roles_hosts are properly prepared
Currently with_dict loops are used to compile oc_roles. This
is inefficient and unnnecessary because we can prepare the
fact using filters instead. This also ensures that the result
is just a list and will not end up being a string unexpectedly.

We also do something similar for the controller_role_name fact.
To ensure like behaviour we pick the last item in the list from
the match.

To ensure that controllers_from_inventory is returned as a list
(which is expected) and not as dict_keys(...) with python3 we
use the list filter. See [a] for more information about this.

Finally, for the oc_roles_hosts fact loop we ensure that the
syntax is updated to work with more up-to-date Ansible.

[a] https://docs.ansible.com/ansible/2.8/user_guide/playbooks_python_version.html#python-version-and-templating
Change-Id: I5cfc593af63db5f885bcda5094340bcc1bf97047
(cherry picked from commit bc52a328bdd080fbf191104cf3b064a4e68d01ac)