Initital ffwd2 patch to automate OS upgrade.
This patch updates some of the ffu existing variables in
the tripleo-upgrade role to be able to upgrade to the
ffu (Queens to Train).

Also, it includes the undercloud operating system upgrade
logic as part of the upgrade of the undercloud (which will
be used mostly downstream). It leverages from the existing
play in common/undercloud_os_upgrade.yaml and adds the
specific tasks right before invoking it.

As part of the refactoring, the fast forward upgrades
automation tasks also make use of the imporved workarounds
mechanism, so all the operations (update,upgrade,ffwd)
share the same format for the workarounds.

Removing too the python-jsonpointer package before running
leapp upgrade on the undercloud node, as it is conflicting
after the upgrade and it does seem to not impact into the
leapp or subscription-manager packages dependencies.

Change-Id: I3b917434b9bc7d5d288b702e85a07b2356fa6896