Ensure that ceph-ansible is installed in the undercloud.
This change impacts mostly to the upgrade from OSP14 to OSP15
in which we clean up all the openstack packages, however it's
also a nice check for the Undercloud upgrade. When making sure
we have the latest content on tht,tc,tripleoclient, etc... we
also ensure that ceph-ansible is installed if ceph is being
deployed.

Change-Id: I205425f31412c856e754539f67b039215a3b2189