Improve ceph services discovery.
In the upgrade from Rocky to Stein we are required to run
an extra step to migrate the ceph systemd unit files to
work with podman instead of Docker. To know which roles
need this step to be run, this patch improves a little
bit the logic by storing in a enabled_services fact all
the services enabled per role. This allows us to check
later, when creating the templates, if the role we are
creating the templates for has ceph enabled or not.

This change is also useful for the FFWD upgrade from 13
to 16, as the workflow will be almost the same.

Change-Id: I5f4a92587ae5862219f3949bff2f83debf634c5c