Adapt FFWD2 Undercloud upgrade workflow.
In theory, in this FFWD iteration (Queens to Train) it won't
be needed to perform an Undercloud upgrade sequentially. Therefore
we don't need to build the same script to upgrade the undercloud
three times. Instead, we will make use of the very same script
used to upgrade the undercloud in updates and upgrades.

This patch also removes some unnecessary playbooks from the last
FFWD and splits the scripts creation into undercloud and overcloud
scripts creation plays and moves the configure_uc_containers.yaml
playbook into the common/ folder so it can be used by the
fast-forward-upgrade and upgrades workflows.

Change-Id: I975fc0c5bf25681bb944415eaccd3a0672998d01