Modify overcloud upgrade run loop.
The way the current playbook in charge of upgrading the
overcloud nodes works is by iterating over all the roles
registered in oc_roles omiting any failure that could
appear.  It's then, when the upgrade for all the roles
has finished, when we check the state of each upgrade.

This approach keeps the HW resource reserved for a very
long time in those cases in which we fail in the very
first Controllers upgrade. This is not a sustainable
thing taking into account the limited resources we have
in the CI.

This approach will iterate over each role, running the
overcloud upgrade run command for each but checking the
state of the upgrade once it finishes. If the first role
upgrade fails we won't move on into the next.

Change-Id: Ic7de17bbcaacd9902d13461dfc4ad4ffa01e0b57