Add needed parameters in containers-prepare-parameters for FFWD2.
We need to include few parameters in the containers-prepare-parameters.yaml
file to be able to perform the upgrade from Queens to Train, one set is the
ceph3_namespace parameters and the other one the Stein namespace ones. So
far they are being hardcoded in the parameters default, so it is possible
to override them if needed.

Change-Id: I4601e9d32ea3cd29a8f80189984e0259d3f29cf4
(cherry picked from commit fb48a1fe68e1542e4191615e79abd057a8864495)