FFWD2 Overcloud upgrade automation.
This is a first draft on the automated workflow to cover the
overcloud upgrade from Queens to Train. A lot of code has been
restored from the upgrade workflow from Rocky to Stein.

In summary, this change will create the overcloud scripts needed
for each of the role hosts (the upgrade is now done node by node)
and change the logic to be able to execute the upgrade for each
of these nodes. To get the nodes per role it relies on the
oc_roles_hosts, a dictionary with all the roles as index and the
hosts (ordered) as values.

In an iteration, a further cleaning up of the all FFU remaining
code will be done.

Change-Id: Idd0bb509c6673e96859fc479b08d33317d9249ea
(cherry picked from commit 5a17cdc6afc42d36634865efbebd05f0b9449fa1)