Add new rhsm_overcloud_env parameter to tripleo-upgrade.
In order to simplify the way we could pass the rhsm configuration
files during the upgrade. This patch adds a new parameter in which
the location of the rhsm credentials for the overcloud will be
stored. When rhsm_overcloud_env is passed (and different than empty
string) then, the upgrade/update prepare steps will include the
tht default rhsm.yaml environment file and the location passed in
rhsm_overcloud_env.

At the same time, it enables a new infrared option to facilitate
the parameter passing when calling to tripleo-upgrade as an infrared
plugin.

Change-Id: I29d1a3ef05e3313360cacf2b17bc0b519852029a