Get rid of the old container images environment file.
For more than three releases, the way to configure the container
images is by using the ContainerImagePrepare service (via the
container-prepare-parameters.yaml file). However, the update
implementation still kept a reference to the old environment
file which contained all the references to the container images.
This file needs to be removed completely from the update prepare
and converge templates, as if the content isn't updated then
the update will point to the old images. This way we rely only
on the container-prepare-parameter.yaml content.

Change-Id: I11767c9c0c398f479ded14db1163693c6e87f621