FFWD automation for HA and composable environments
We need to run in quite tedious order to upgrade properly
the Pacemaker managed roles. First we split the environment
into Pacemaker enabled, Non-Pacemaker Non-Compute and Compute
groups of roles:

Pacemaker: ['Controller','Database','Messaging']
Non-Pacemaker&Non-Compute: ['Ceph','Networking','Objectstorage']
Compute: ['Compute','OSDCompute']

For Pacemaker we have to upgrade one of each roles in the Pacemaker
group together - controller-0,database-0,messaging-0 to ensure that
upgrade will work. We can run system upgrade in parallel on them
too.

The tricky part with Pacemaker is that than we need to run upgrade
on the one node of each role in the Pacemaker group but also on all
previous Pacemaker nodes to ensure we don't create several separated
clusters.

For the rest it's simply faster to run for each role in the group.

Change-Id: I65b3f3f45f975be90438b3cdc3c8e24cd290d0a8