Move inventory filters to tripleo-upgrade
The inventory filters for tripleo-upgrade were originally merged to
tripleo-ansible to promote re-use however we want to avoid adding this
as a dependency for tripleo-upgrade as it is not necessarily running on
the undercloud where we can assume tripleo-ansible is installed.
The use case for these filters is quite obscure and specific to the
tripleo-upgade playbooks (manually parsing the inventory yaml in an
ansible task) so the easiest option is to move them to tripleo-upgrade.

Change-Id: I712f752c6a8258e20bd0f73453ce452beb16509b